+++ packages/adaptive_scaffold/lib/src/adaptive_scaffold.dart
@@ -3,7 +3,6 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:masonry_grid/masonry_grid.dart';
 
 import 'adaptive_layout.dart';
 import 'breakpoints.dart';
@@ -309,15 +308,19 @@ class AdaptiveScaffold extends StatefulWidget {
       int currentIndex = 0,
       double iconSize = 24}) {
     return Builder(
-        builder: (_) => BottomNavigationBar(
-            currentIndex: currentIndex,
-            iconSize: iconSize,
-            items: destinations
-                .map((NavigationDestination e) => _toBottomNavItem(e))
-                .toList()));
+      builder: (_) {
+        return BottomNavigationBar(
+          currentIndex: currentIndex,
+          iconSize: iconSize,
+          items: destinations
+              .map((NavigationDestination e) => _toBottomNavItem(e))
+              .toList(),
+        );
+      },
+    );
   }
 
-  /// Public helper method to be used for creating a [MasonryGrid] following m3
+  /// Public helper method to be used for creating a staggered grid following m3
   /// specs from a list of [Widget]s
   static Builder toMaterialGrid({
     List<Widget> thisWidgets = const <Widget>[],
@@ -361,10 +364,11 @@ class AdaptiveScaffold extends StatefulWidget {
           SliverToBoxAdapter(
             child: Padding(
               padding: EdgeInsets.all(thisMargin),
-              child: MasonryGrid(
-                column: itemColumns,
-                crossAxisSpacing: kMaterialGutterValue,
-                mainAxisSpacing: kMaterialGutterValue,
+              child: _BrickLayout(
+                columns: itemColumns,
+                columnSpacing: kMaterialGutterValue,
+                itemPadding:
+                    const EdgeInsets.only(bottom: kMaterialGutterValue),
                 children: thisWidgets,
               ),
             ),
@@ -441,7 +445,9 @@ class AdaptiveScaffold extends StatefulWidget {
   static Widget fadeOut(Widget child, Animation<double> animation) {
     return FadeTransition(
       opacity: CurvedAnimation(
-          parent: ReverseAnimation(animation), curve: Curves.easeInCubic),
+        parent: ReverseAnimation(animation),
+        curve: Curves.easeInCubic,
+      ),
       child: child,
     );
   }
@@ -614,4 +620,103 @@ BottomNavigationBarItem _toBottomNavItem(NavigationDestination destination) {
     label: destination.label,
     icon: destination.icon,
   );
+}
+
+class _BrickLayout extends StatelessWidget {
+  const _BrickLayout({
+    this.columns = 1,
+    this.itemPadding = EdgeInsets.zero,
+    this.columnSpacing = 0,
+    required this.children,
+  });
+
+  final int columns;
+  final double columnSpacing;
+  final EdgeInsetsGeometry itemPadding;
+  final List<Widget> children;
+
+  @override
+  Widget build(BuildContext context) {
+    int i = -1;
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        Expanded(
+          child: CustomMultiChildLayout(
+            delegate: _BrickLayoutDelegate(
+              columns: columns,
+              columnSpacing: columnSpacing,
+              itemPadding: itemPadding,
+            ),
+            children: children
+                .map<Widget>(
+                  (Widget child) => LayoutId(id: i += 1, child: child),
+                )
+                .toList(),
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class _BrickLayoutDelegate extends MultiChildLayoutDelegate {
+  _BrickLayoutDelegate({
+    this.columns = 1,
+    this.columnSpacing = 0,
+    this.itemPadding = EdgeInsets.zero,
+  });
+
+  final int columns;
+  final EdgeInsetsGeometry itemPadding;
+  final double columnSpacing;
+
+  @override
+  void performLayout(Size size) {
+    final BoxConstraints looseConstraints = BoxConstraints.loose(size);
+    final BoxConstraints fullWidthConstraints =
+        looseConstraints.tighten(width: size.width);
+
+    final List<Size> childSizes = <Size>[];
+    int childCount = 0;
+    // Count how many children we have.
+    for (; hasChild(childCount); childCount += 1) {}
+    final BoxConstraints itemConstraints = BoxConstraints(
+      maxWidth: fullWidthConstraints.maxWidth / columns -
+          columnSpacing / 2 -
+          itemPadding.horizontal,
+    );
+
+    for (int i = 0; i < childCount; i += 1) {
+      childSizes.add(layoutChild(i, itemConstraints));
+    }
+
+    int columnIndex = 0;
+    int childId = 0;
+    final double totalColumnSpacing = columnSpacing * (columns - 1);
+    final double columnWidth = (size.width - totalColumnSpacing) / columns;
+    final double topPadding = itemPadding.resolve(TextDirection.ltr).top;
+    final List<double> columnUsage =
+        List<double>.generate(columns, (int index) => topPadding);
+    for (final Size childSize in childSizes) {
+      positionChild(
+        childId,
+        Offset(
+          columnSpacing * columnIndex +
+              columnWidth * columnIndex +
+              (columnWidth - childSize.width) / 2,
+          columnUsage[columnIndex],
+        ),
+      );
+      columnUsage[columnIndex] += childSize.height + itemPadding.vertical;
+      columnIndex = (columnIndex + 1) % columns;
+      childId += 1;
+    }
+  }
+
+  @override
+  bool shouldRelayout(_BrickLayoutDelegate oldDelegate) {
+    return itemPadding != oldDelegate.itemPadding ||
+        columnSpacing != oldDelegate.columnSpacing;
+  }
 }