+++ packages/pigeon/platform_tests/macos_swift_unit_tests/pigeons/messages.dart
@@ -1,10 +1,10 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-package dev.flutter.packages.adaptive_scaffold.example
 
-import io.flutter.embedding.android.FlutterActivity
+import 'package:pigeon/pigeon.dart';
 
-class MainActivity: FlutterActivity() {
-
+@HostApi()
+abstract class Api {
+  String getPlatform();
 }