+++ packages/pigeon/run_tests.sh
@@ -191,6 +191,10 @@ run_mock_handler_tests() {
   dart run tool/run_tests.dart -t mock_handler_tests
 }
 
+run_macos_swift_unittests() {
+  dart run tool/run_tests.dart -t mac_swift_unittests
+}
+
 run_dart_compilation_tests() {
   local temp_dir=$(mktmpdir)
   local flutter_project_dir=$temp_dir/project
@@ -311,6 +315,7 @@ should_run_flutter_unittests=true
 should_run_ios_e2e_tests=true
 should_run_ios_unittests=true
 should_run_mock_handler_tests=true
+should_run_macos_swift_unittests=true
 while getopts "t:l?h" opt; do
   case $opt in
   t)
@@ -321,6 +326,7 @@ while getopts "t:l?h" opt; do
     should_run_ios_e2e_tests=false
     should_run_ios_unittests=false
     should_run_mock_handler_tests=false
+    should_run_macos_swift_unittests=false
     case $OPTARG in
     android_unittests) should_run_android_unittests=true ;;
     dart_compilation_tests) should_run_dart_compilation_tests=true ;;
@@ -329,6 +335,7 @@ while getopts "t:l?h" opt; do
     ios_e2e_tests) should_run_ios_e2e_tests=true ;;
     ios_unittests) should_run_ios_unittests=true ;;
     mock_handler_tests) should_run_mock_handler_tests=true ;;
+    macos_swift_unittests) should_run_macos_swift_unittests=true ;;
     *)
       echo "unrecognized test: $OPTARG"
       exit 1
@@ -344,6 +351,7 @@ while getopts "t:l?h" opt; do
   ios_e2e_tests          - End-to-end objc tests run on iOS Simulator
   ios_unittests          - Unit tests on generated Objc code.
   mock_handler_tests     - Unit tests on generated Dart mock handler code.
+  macos_swift_unittests  - Unit tests on generated Swift code on macOS.
   "
     exit 1
     ;;
@@ -389,4 +397,7 @@ if [ "$should_run_ios_e2e_tests" = true ]; then
 fi
 if [ "$should_run_android_unittests" = true ]; then
   run_android_unittests
+fi
+if [ "$should_run_macos_swift_unittests" = true ]; then
+  run_macos_swift_unittests
 fi