+++ packages/flutter_markdown/CHANGELOG.md
@@ -4,6 +4,11 @@
   Consider using `OrderedListWithCheckBoxSyntax` or `UnorderedListWithCheckBoxSyntax` as a replacement.
 * Changes `_buildCheckbox()` to inspect state of checkbox input element by existence of `'checked'` attribute.
 
+## 0.6.10+6
+
+* Removes print logged when not handling hr for alignment.
+* Removes print logged when not handling li for alignment.
+
 ## 0.6.10+5
 
 * Fixes lint warnings.